Fix bazel build for external workspaces; accept C++11 time types by default during a C++11 build.
Rename bazel/BUILD to bazel/BUILD.bazel.
This makes it clear that bazel/ is not really a distinct package.
The file bazel/BUILD.bazel is just a copy of the default bazel BUILD file.

Introduce (in bazel/pkg_path_name.bzl) a function
that returns the pathname of the calling package,
This is used to construct include paths in the BUILD file.
Previously, the file used just PACKAGE_NAME, which
did not work for external packages in bazel.

Change the rules "nsync" and "nsync_cpp" in the BUILD file
to specify
    includes = ["."],
which allows rules that transitively depend on these rules to include the
public headers.

In public/nsync_time_internal.h, allow the clients to use C++11 time types
for timeouts with the wait functions if the module is compiled with C++11,
either on Windows or other compilers.   Previously, this was enabled only
if the compiler defined NSYNC_USE_CPP11_TIMEPOINT.