Allow nsync to work with bazel on MacOS; fix C++11 time_point use on MacOS.
On MacOS, the cpu type name for bazel is "darwin";
fix this in both the BUILD file and bazel/BUILD.bazel.

Also fix the use of time_point and duration for C++11.  On MacOS, this has to
be done carefully, because a duration with one ratio is not type compatible
with a duration with another ratio---and you can't predict what the duration
ratio will be for system_clock on any given platform.