Bug fix for nsync_cv transfer of waiters to the mutex.
When a condition variable wakeup occurs and the associated mutex is held, the
intent is that waiters are transferred to the mutex waiter queue rather than
being woken only to block anew on the mutex.

There were at least three bugs in this code:
- I used the wrong condition for "just one waiter in the list".
  My expression used the condition for a raw circular list (first_waiter !=
  next), but the API to the list returns NULL when asked for the next element
  from the last, so it should be (next != NULL).
- Setting the designated waker bit is more subtle here than I assumed;
  the easiest fix is to forgo the opimization in this path.  After all,
  condition variable wakeup implies that some thread had blocked,
  which is a fairly heavyweight operation, so the optmization is not
  worth as much as a fraction.
- Transfer to the mutex queue is also subtle if the mutex was not held by some
  thread.  It may be possible to perform such a transfer once the other
  two issues are fixed, perhaps by checking that there is already a designated
  waker, but it makes the code harder to think about.  So I forgo this
  optimization also.
Besides subtlety, another reason to avoid heroic optimizations in the case
where the mutex is not held is that I wouldn't recommend that anyone normally
signal with the mutex not held anyway, since it's unsafe in some cases (e.g.,
imagine the object containing the mutex is reference counted, and the count is
protected by that mutex, so the object may be freed by the time of the signal).
That bias in expectations is part of why I missed this in testing in the first
place.

This change fixes the queue length test, allows transfer only if the mutex is
actually held, and does not attempt the optimization of setting the designated
waker bit.