Rename BUILD.bazel to bazel_BUILD to prevent bazel using it as a BUILD file.
I had used the name "BUILD.bazel" for the name of an inactive copy of a bazel
BUILD file that should not be interpreted as a BUILD file in its existing
directory.  (The file is copied to BUILD at the root of the nsync tree to make
an open source, bazel-buildable version of the tree.)

Alas, bazel will interpret "BUILD.bazel" as though it were "BUILD" because:
1) when using the "makefile" build,
      tensorflow/contrib/makefile/download_dependencies.sh
   puts a copy of the nsync tree inside the tensorflow tree
   (under tensorflow/contrib/makefile/downloads), and
2) various build rules in the tensorflow tree use recursive glob syntax to
   refer to entire trees, and some thus include everything within
   tensorflow/contrib/makefile/downloads.
As a result, if someone first uses download_dependencies.sh, and then invokes
bazel on a rule that uses a recursive glob that covers the downsloads
directory, bazel will parse all the BUILD files within all the downloaded
packages(!), and thus fail if any contain BUILD (or BUILD.bazel) files not
valid in their contexts.

This change renames BUILD.bazel to bazel_BUILD.  Bazel doesn't seem
to recognize that name yet.