Fix potential undefined behaviour found with -fsanitize=function.
In two source files I was calling functions whose pointers have type
    void (*) (nsync_mu *)
via function pointers with type
    void (*) (void *)

While that will work on almost all implementations,
it is not required to work by the C and C++ standards.

This change introduces local functions to do the conversion in a safe way
(the conversion is needed to implement the generic CV wait operation).

The new code involves an extra function call, but this is not on a fast path:
it happens when waiting on a condition variable, which will usually involve a
trip through the kernel.