Bazel and cmake support for FreeBSD in nsync.
Bazel BUILD file
- add select() clausesx for FreeBSD.

CMakeLists.txt
- fix the FreeBSD source list.  Also fix for NetBSD and OpenBSD; they are the same.

Add a Makefile for C++11 build on FreeBSD.

Usually, the tests can be linked by invoking the compiler as though it were the
linker.  But with clang on FreeBSD, the "-x c++" used to tell the compiler to
treat the .c files as C++ also causes .o files to be treated as C++ source
files.  So this change also modifies make.common to make it possible to use a
different command for the linker.

Fix a bug in mkdep.c, which would not recognize the suffix .CC as C++.