Use futex to implement low-level semphore on Linux, even when building in portable C++11 mode.
Previously in C++11 builds in Linux, I had used a C++11 condition variable to
implement the semaphore that nsync uses to block.
However, for some reason the condition variable in Linux' C++11 library is
extremely and surprisingly slow:  perhaps ten times slower than the basic
operating system primitive.

This change uses the underlying primitive (futex) directly,
to avoid this performance problem.

This is fixed in the Makefile, cmake and bazel builds.