CMake: version sofiles and respect libdir
Version the libraries like libnsync.so.1.20.0 and set the symlinks
properly. The version is taken from the VERSION file in the root, it
should be updated with new releases. The minimum is kept CMake version
is kept at 2.8 but >=3.0 is strongly recommended. Since the minimum is
<3.0, policy CMP0048 must be set to NEW.

64bit distros use /usr/lib64 not /usr/lib so this needs to be
configurable as well.

The source properties for C++ are set outside of enabling the tests
otherwise building the C++ library with tests disabled fails.

Signed-off-by: Jason Zaman <jason@perfinion.com>