Patch 1 (#7)
* Update BUILD

Our team at Arm Inc. tested Tensorflow that has a dependency on nsync. We found that the cpu type has to be "aarch64" instead of "arm" for correct compilation on aarch64 servers.

* Update bazel_BUILD

Our team at Arm Inc. tested Tensorflow that has a dependency on nsync. We found that the cpu type has to be "aarch64" instead of "arm" for correct compilation on aarch64 servers.