Introduce an implementation of nsync's internal semaphore that uses MacOS Grand Central Dispatch primitives. For documentation of that primitive, see:     https://developer.apple.com/documentation/dispatch/1452955-dispatch_semaphore_create     https://developer.apple.com/documentation/dispatch/1453087-dispatch_semaphore_wait     https://developer.apple.com/documentation/dispatch/1452919-dispatch_semaphore_signal
This seamaphore is faster for wakeups on the Mac than the implementation using
the pthread API.  For example, for the ping_pong benchmark (which just does
waits and wakeups back and forth between threads) for mu_wait and
cv_wait:
   Benchmark pingpong_test                                  ops     time   ops/sec  time/op
   ----------------------------------------------------------------------------------------
pthread implementation:
   benchmark_ping_pong_mu                                359375    3.66s   9.8e+04  10.17us
   benchmark_ping_pong_mu_cv                             359375    3.83s   9.4e+04  10.66us
gcd_implementation:
   benchmark_ping_pong_mu                                359375    2.54s   1.4e+05   7.08us
   benchmark_ping_pong_mu_cv                             359375    2.51s   1.4e+05   6.98us

Create a Makefile that uses the new semaphore implementation on the Mac, for testing.

In cv_test.c and mu_wait_test.c, modify the way values are converted between
pointers and integers for the producer/consumer performance test to avoid
warning on some compilers about pointer arithmetic on the nil pointer.