Suppress compiler warnings on newer compilers that have additional checks.
GCC 8.0 in C++ mode introduced a warning if memset() is used to zero a struct,
even if it POD.  This causes the compiler to be chatty if nsync is build using g++.
Most of the changes are to cast the pointer argument to memset() to "void *" to avoid
the warning.

GCC 8.0 in C++ mode also started warning if a C-style cast changed the
"constness" of the data.  I'd accidentally done this in testing/testing.c; now fixed.

NetBSD has started using C99-style designated initializers in the standard pthread
initializer constants even if using C89; clang 8.0 warns about this.  This
change turns the warning off in the netbsd/clang Makefiles.