In nsync, avoid an ABSL depepdency for TSAN annotations.
The code used to use ABSL macros for TSAN annotations that suppress warnings.

This was a problem because:
- ABSL is changing the names of the macros(!), and the team seems intent on
  forcing clients to make the incompatible change, even if the users of those
  clients have not yet switched to a suifficiently recent version of ABSL.  And
  they are doing this despite their (apparently false) claim of a 5-year API
  stability guarantee.
- The code had hard-wired a header file path that was wrong for most users,
  effectively preventing them from running with TSAN.
- The conditional compilation to detect TSAN worked only for clang, and not for
  gcc, even though gcc supports TSAN.

This change drops the ABSL dependency so there's no header file to include;
it uses conditional compilation that detects TSAN on moth gcc and clang; and
it uses the gcc/clang annotation routines directly.

The version number is also increased.