Fix CMake builds using MSVC's Clang frontend
Currently, nsync suppresses a number of MSVC warnings in order to keep
the build-log size manageable.  The warning suppressions are added if
the operating system is detected as Windows, which is no longer quite
correct - MSVC has a Clang frontend now that does not understand the
"/wdXXXX" switch format.  Instead, it treats those flags as files to
compile.

This commit fixes the issue by defining those warning suppressions if
the compiler ID is MSVC, which is the correct check.

It also switches from "add_definitions" to "add_compile_options", which
is a little more correct.  The scoping rules are the same, but
"add_definitions" is intended more for preprocessor definitions.